Renaming fetch-utils.ts to fetch-tuils.server.ts
