+++ frontend/app/services/application-status-service.server.ts
@@ -4,7 +4,7 @@ import { z } from 'zod';
 import { getAuditService } from '~/services/audit-service.server';
 import { getInstrumentationService } from '~/services/instrumentation-service.server';
 import { getEnv } from '~/utils/env.server';
-import { getFetchFn } from '~/utils/fetch-utils';
+import { getFetchFn } from '~/utils/fetch-utils.server';
 import { getLogger } from '~/utils/logging.server';
 
 const log = getLogger('application-status-service.server');
