+++ frontend/app/services/benefit-application-service.server.ts
@@ -7,7 +7,7 @@ import { BenefitApplicationRequest, benefitApplicationRequestSchema, benefitAppl
 import { getAuditService } from '~/services/audit-service.server';
 import { getInstrumentationService } from '~/services/instrumentation-service.server';
 import { getEnv } from '~/utils/env.server';
-import { getFetchFn, instrumentedFetch } from '~/utils/fetch-utils';
+import { getFetchFn, instrumentedFetch } from '~/utils/fetch-utils.server';
 import { getLogger } from '~/utils/logging.server';
 
 const log = getLogger('benefit-application-service.server');
